+++ __main__.py
@@ -5,4 +5,4 @@
 from send.send import main
 
 if __name__ == "__main__":
-    main()
+    raise SystemExit(main())