+++ send/__main__.py
@@ -1,0 +1,11 @@
+"""
+Module entry point for Auto Announcements.
+Copyright (C) 2017-2026 willtheorangeguy
+"""
+
+# pylint: disable=import-error
+
+from send.send import main
+
+if __name__ == "__main__":
+    raise SystemExit(main())